+++ source/maker-registry/scripts/deploy.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-console */
 const fs = require('fs')
+const Confirm = require('prompt-confirm')
 const { ethers, run } = require('hardhat')
 const { chainNames, stakingTokenAddresses } = require('@airswap/constants')
 const registryDeploys = require('../deploys.js')
@@ -15,27 +16,32 @@ async function main() {
   const tokenCost = 0
 
   console.log(`Deploying on ${chainNames[chainId].toUpperCase()}`)
-  const registryFactory = await ethers.getContractFactory('MakerRegistry')
-  const registryContract = await registryFactory.deploy(
-    stakingToken,
-    obligationCost,
-    tokenCost
-  )
-  await registryContract.deployed()
-  console.log(`Deployed: ${registryContract.address}`)
+  console.log(`Staking token: ${stakingToken}`)
 
-  registryDeploys[chainId] = registryContract.address
-  fs.writeFileSync(
-    './deploys.js',
-    `module.exports = ${JSON.stringify(registryDeploys, null, '\t')}`
-  )
-  console.log('Updated deploys.js')
+  const prompt = new Confirm('Proceed to deploy?')
+  if (await prompt.run()) {
+    const registryFactory = await ethers.getContractFactory('MakerRegistry')
+    const registryContract = await registryFactory.deploy(
+      stakingToken,
+      obligationCost,
+      tokenCost
+    )
+    await registryContract.deployed()
+    console.log(`Deployed: ${registryContract.address}`)
 
-  console.log(
-    `\nVerify with "yarn verify --network ${chainNames[
-      chainId
-    ].toLowerCase()}"\n`
-  )
+    registryDeploys[chainId] = registryContract.address
+    fs.writeFileSync(
+      './deploys.js',
+      `module.exports = ${JSON.stringify(registryDeploys, null, '\t')}`
+    )
+    console.log('Updated deploys.js')
+
+    console.log(
+      `\nVerify with "yarn verify --network ${chainNames[
+        chainId
+      ].toLowerCase()}"\n`
+    )
+  }
 }
 
 main()
