+++ source/staking/scripts/deploy.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-console */
 const fs = require('fs')
+const Confirm = require('prompt-confirm')
 const { ethers, run } = require('hardhat')
 const { chainNames, stakingTokenAddresses } = require('@airswap/constants')
 const stakingDeploys = require('../deploys.js')
@@ -19,29 +20,32 @@ async function main() {
   console.log(`Deploying on ${chainNames[chainId].toUpperCase()}`)
   console.log(`Staking token: ${stakingToken}`)
 
-  const stakingFactory = await ethers.getContractFactory('Staking')
-  const stakingContract = await stakingFactory.deploy(
-    stakingToken,
-    name,
-    symbol,
-    duration,
-    minDelay
-  )
-  await stakingContract.deployed()
-  console.log(`Deployed: ${stakingContract.address}`)
+  const prompt = new Confirm('Proceed to deploy?')
+  if (await prompt.run()) {
+    const stakingFactory = await ethers.getContractFactory('Staking')
+    const stakingContract = await stakingFactory.deploy(
+      stakingToken,
+      name,
+      symbol,
+      duration,
+      minDelay
+    )
+    await stakingContract.deployed()
+    console.log(`Deployed: ${stakingContract.address}`)
 
-  stakingDeploys[chainId] = stakingContract.address
-  fs.writeFileSync(
-    './deploys.js',
-    `module.exports = ${JSON.stringify(stakingDeploys, null, '\t')}`
-  )
-  console.log('Updated deploys.js')
+    stakingDeploys[chainId] = stakingContract.address
+    fs.writeFileSync(
+      './deploys.js',
+      `module.exports = ${JSON.stringify(stakingDeploys, null, '\t')}`
+    )
+    console.log('Updated deploys.js')
 
-  console.log(
-    `\nVerify with "yarn verify --network ${chainNames[
-      chainId
-    ].toLowerCase()}"\n`
-  )
+    console.log(
+      `\nVerify with "yarn verify --network ${chainNames[
+        chainId
+      ].toLowerCase()}"\n`
+    )
+  }
 }
 
 main()
