+++ source/wrapper/scripts/deploy.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-console */
 const fs = require('fs')
+const Confirm = require('prompt-confirm')
 const { ethers, run } = require('hardhat')
 const swapDeploys = require('@airswap/swap-erc20/deploys.js')
 const wrapperDeploys = require('../deploys.js')
@@ -18,26 +19,29 @@ async function main() {
   console.log(`Swap: ${swapAddress}`)
   console.log(`Wrapped: ${wrappedTokenAddress}`)
 
-  const wrapperFactory = await ethers.getContractFactory('Wrapper')
-  const wrapperContract = await wrapperFactory.deploy(
-    swapAddress,
-    wrappedTokenAddress
-  )
-  await wrapperContract.deployed()
-  console.log(`Deployed: ${wrapperContract.address}`)
+  const prompt = new Confirm('Proceed to deploy?')
+  if (await prompt.run()) {
+    const wrapperFactory = await ethers.getContractFactory('Wrapper')
+    const wrapperContract = await wrapperFactory.deploy(
+      swapAddress,
+      wrappedTokenAddress
+    )
+    await wrapperContract.deployed()
+    console.log(`Deployed: ${wrapperContract.address}`)
 
-  wrapperDeploys[chainId] = wrapperContract.address
-  fs.writeFileSync(
-    './deploys.js',
-    `module.exports = ${JSON.stringify(wrapperDeploys, null, '\t')}`
-  )
-  console.log('Updated deploys.js')
+    wrapperDeploys[chainId] = wrapperContract.address
+    fs.writeFileSync(
+      './deploys.js',
+      `module.exports = ${JSON.stringify(wrapperDeploys, null, '\t')}`
+    )
+    console.log('Updated deploys.js')
 
-  console.log(
-    `\nVerify with "yarn verify --network ${chainNames[
-      chainId
-    ].toLowerCase()}"\n`
-  )
+    console.log(
+      `\nVerify with "yarn verify --network ${chainNames[
+        chainId
+      ].toLowerCase()}"\n`
+    )
+  }
 }
 
 main()
